refactor: replace id of Integer type with Long type
(cherry picked from commit 21b5935778eac455954d5616532caf6e2bdfb8c6)

+++ real-estate/src/main/java/org/apartment/controller/PropertyController.java
@@ -28,7 +28,7 @@ public class PropertyController {
       @RequestPart("data") @Valid PropertyDto propertyDto,
       @RequestParam("image") MultipartFile[] imageFiles, @RequestHeader("X-User-Id") String userId)
       throws Exception {
-    propertyDto.setPostedBy(Integer.parseInt(userId));
+    propertyDto.setPostedBy(Long.parseLong(userId));
     Property property = PropertyMapper.INSTANCE.toEntity(propertyDto);
     Property createdProperty = propertyService.createProperty(property, imageFiles);
     return ResponseEntity.ok(createdProperty);
